Add authotization to some enitites add fixing some bugs

+++ Controllers/OutgoesController.cs
@@ -3,11 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using CConstsProject.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace CCostsProject.Controllers
 {
+    [Authorize(AuthenticationSchemes = "Bearer")]
     [Route("api/[controller]")]
     public class OutgoesController : Controller
     {
@@ -18,23 +20,25 @@ namespace CCostsProject.Controllers
             db = context;
             Worker = new DbWorker(db);
         }
-        
+        [Authorize]
         [HttpPost("AddOutgo")]
         public  IActionResult Post([FromBody]Outgo outgo)
         {
             if (outgo != null)
             {
+                outgo.User = db.Users.FirstOrDefault(u => u.UserName == User.Identity.Name);
                 Worker.AddOutgo(outgo);
-               
-                return Ok(outgo);
+                Worker.MakeOutgo(User.Identity.Name,outgo.Money);
+                return Ok(outgo+"current cash:");
             }
             return Forbid();
         }
+        [Authorize]
         [HttpPost("EditOutgo")]
         public IActionResult Post(int id,double Money,DateTime Date)
         {
             Outgo outgo = db.Outgos.FirstOrDefault(o => o.Id == id);
-            if (outgo != null)
+            if (outgo != null&&outgo.User.UserName==User.Identity.Name)
             {
                 outgo.Money = Money;
                 outgo.Date = Date;
@@ -43,11 +47,12 @@ namespace CCostsProject.Controllers
             }
             return Forbid();
         }
+        [Authorize]
         [HttpDelete("DelelteOutgo")]
         public IActionResult Delete(int id)
         {
             Outgo outgo = db.Outgos.FirstOrDefault(o => o.Id == id);
-            if (outgo != null)
+            if (outgo != null && outgo.User.UserName == User.Identity.Name)
             {
                 db.Outgos.Remove(outgo);
                 db.SaveChanges();
@@ -55,10 +60,22 @@ namespace CCostsProject.Controllers
             }
             return Forbid();
         }
+        [Authorize]
+        [HttpGet("{id}")]
+        public IActionResult Get(int id)
+        {
+            Outgo outgo = Worker.GetOutgo(id);
+            if (outgo != null)
+            {
+                return Json(outgo);
+            }
+            return NotFound();
+        }
+        [Authorize]
         [HttpGet("GetOutgoes")]
         public IActionResult GetOutgoes()
         {
-            return new JsonResult(db.Outgos.Include(o=>o.Item).Include(o=>o.User).ToList());
+            return new JsonResult(db.Outgos.Include(o=>o.Item).Include(o=>o.User).Where(o=>o.User.UserName==User.Identity.Name || o.User.Family == Worker.GetFamilyByUserName(User.Identity.Name)).ToList());
         }
     }
 }
